+++ Examples/1_LoRa_Sensor/1_LoRa_Sensor.ino
@@ -8,23 +8,20 @@
 
 
 #include "fdrs_sensor.h"
-#include "sensor_setup.h"
-
-FDRSLoRa FDRS(GTWY_MAC,READING_ID,MISO,MOSI,SCK,SS,RST,DIO0,BAND,SF);
 
 float data1;
 float data2;
 
 void setup() {
-  FDRS.begin();
+  beginFDRS();
 }
 void loop() {
   data1 = readHum();
-  FDRS.load(data1, HUMIDITY_T);
+  loadFDRS(data1, HUMIDITY_T);
   data2 = readTemp();
-  FDRS.load(data2, TEMP_T);
-  FDRS.send();
-  FDRS.sleep(10);  //Sleep time in seconds
+  loadFDRS(data2, TEMP_T);
+  sendFDRS();
+  sleepFDRS(10);  //Sleep time in seconds
 }
 
 float readTemp() {
