--- Examples/FDRS_Install/fdrs_globals.h
@@ -1,8 +1,0 @@
-#define GLOBAL_SSID "Your SSID"
-#define GLOBAL_PASS "Password"
-#define GLOBAL_MQTT_ADDR "192.168.0.8"
-
-#define GLOBAL_BAND 915E6 //LoRa Frequency Band
-#define GLOBAL_SF 7  //LoRa Spreading Factor
-
-#define GLOBALS
