+++ Examples/Universal_Gateway/fdrs_gateway.h
