+++ FDRS_Gateway/README.md
@@ -45,10 +45,10 @@ Buffers can hold a maximum of 256 DataReadings.
 
 
 
-![Basic](/FDRS_Gateway2000/Basic_Setup.png)
+![Basic](/FDRS_Gateway/Basic_Setup.png)
 
-![Advanced](/FDRS_Gateway2000/Advanced_Setup.png)
+![Advanced](/FDRS_Gateway/Advanced_Setup.png)
 
-![Basic LoRa](/FDRS_Gateway2000/Basic_LoRa_Setup.png)
+![Basic LoRa](/FDRS_Gateway/Basic_LoRa_Setup.png)
 
-![Advanced LoRa](/FDRS_Gateway2000/Advanced_Setup_LoRa.png)
+![Advanced LoRa](/FDRS_Gateway/Advanced_Setup_LoRa.png)
