--- Original_FDRS/Front-End_Modules/FDRS_Blynk/DataReading.h
@@ -1,7 +1,0 @@
-
-typedef struct DataReading {
-  float d;
-  uint16_t id;
-  uint8_t t;
-
-} DataReading;
