+++ Universal_Sensor_beta/sensor_setup.h
@@ -6,11 +6,9 @@
 
 //#include <fdrs_globals.h> //Uncomment when you install the globals file
 
-#define READING_ID    2   //Unique ID for this sensor
-#define GTWY_MAC      0x03 //Address of the nearest gateway
+#define READING_ID    1   //Unique ID for this sensor
+#define GTWY_MAC      0x04 //Address of the nearest gateway
 
-#define USE_ESPNOW
-//#define USE_LORA
 #define DEEP_SLEEP
 //#define POWER_CTRL    14
 #define DEBUG
